Use top_builddir instead of ROOT
Building with MSYS2 was failing due to the command line for linking glib
DLL exceeding the sh.exe limit of 8192 bytes. When sh.exe from MSYS2 was
present in PATH, gmake would use it and run a truncated command.

Using relative filenames shortens that command line to about 5k
regardless of the source directory location.

The use of absolute paths to any files in the build directory should be
strongly discouraged, as the build should not keep memory of the location
where it was built.

Replace ROOT with top_builddir everywhere. The ROOT variable was not
named correctly, the standard name is abs_top_builddir and it remains
available in case it's ever needed.